+++ MediaServer.Sse.Core/Broadcasting/ISseEventBroadcaster.cs
@@ -5,6 +5,8 @@ namespace MediaServer.Sse.Core.Broadcasting;
 
 public interface ISseEventBroadcaster : IDisposable
 {
+    int SubscriberCount { get; }
+
     (Guid Id, ChannelReader<SseEvent> Reader) Subscribe();
 
     void Unsubscribe(Guid id);